Change how key files and AES-GCM IVs work
Modified how AES-GCM IVs work in latchbox so all IVs are unique rather
than random.  Also modified how keyfile data is added to the latchbox
passphrase by using HMAC-SHA512 with the key file as the secret key and
the passphrase as the message rather than appending the SHA512 hash of
the key file to the passphrase.  Updated latchbox-spec.txt.  Also
updated go-runewidth.